requirements.txt
pytest >= 7.4, < 7.5
pyright >= 1.1, < 1.2
ruff >= 0.1, < 0.2

pyinstaller == 6.3.0

openpyxl == 3.1.2
pyyaml == 6.0.2
python-dateutil == 2.8.2
pytz == 2024.1
requests == 2.31.0
